backend/requirements.txt
python-dotenv==1.0.1
scikit-learn==1.3.2
python-docx==1.1.0
pdfminer.six==20221105
email-validator==2.1.0.post1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
sentence-transformers==2.2.2
huggingface_hub==0.20.3
